OffshoreProfitShifting_Replication

Documentation for OffshoreProfitShifting_Replication.

Offshore_Profit_Shifting_Replication.logg

Offshore_Profit_Shifting_Replication.makedataforplot2

Offshore_Profit_Shifting_Replication.makedataforplot7

Offshore_Profit_Shifting_Replication.makedataforplot8

Offshore_Profit_Shifting_Replication.makeplotdata

Offshore_Profit_Shifting_Replication.plots1

Offshore_Profit_Shifting_Replication.plots10

Offshore_Profit_Shifting_Replication.plots11

Offshore_Profit_Shifting_Replication.plots2

Offshore_Profit_Shifting_Replication.plots3

Offshore_Profit_Shifting_Replication.plots4

Offshore_Profit_Shifting_Replication.plots6

Offshore_Profit_Shifting_Replication.plots7

Offshore_Profit_Shifting_Replication.plots8

Offshore_Profit_Shifting_Replication.plots9

Offshore_Profit_Shifting_Replication.process_files

Offshore_Profit_Shifting_Replication.rename_and_process!

Offshore_Profit_Shifting_Replication.table1

Offshore_Profit_Shifting_Replication.table5

Offshore_Profit_Shifting_Replication.table6

Offshore_Profit_Shifting_Replication.logg — Method

logg(x::Vector)

Calculates the logarithmic growth rate for a vector x. It does this by computing the log of the ratio of subsequent elements in the vector. The first element of the resulting vector is NaN as there's no previous element for the first value in x.

Arguments

x::Vector: A numeric vector for which the logarithmic growth rate needs to be computed.

Returns

Vector: A vector of the same length as x, with the first element being NaN and the rest being the log growth rate of x.

Example

```julia julia> logg([1, 2, 3, 4]) 3-element Array{Float64,1}:

Offshore_Profit_Shifting_Replication.makedataforplot2 — Method

makedataforplot2()

This function reads and processes data for creating plot 2. It reads data from the "OutputAggStdError.xlsx" file, performs necessary data manipulations and calculations.

Specifically, the function does the following:

Reads the "OutputAggStdError.xlsx" file.

Creates a DataFrame with all combinations of unique Panels and Years.

Interpolates missing values within the DataFrame.

Joins the DataFrame with the public DataFrame.

Computes statistics across panels for each Year.

Usage

```julia panel_stats = makedataforplot2()

Returns

panel_stats: A DataFrame containing computed statistics for each Panel. This function takes no arguments.

Offshore_Profit_Shifting_Replication.makedataforplot7 — Method

makedataforplot7()

This function prepares data needed for Plot 7. It performs calculations on several variables such as gdibusn, gdibusr, uprod_log, uprod_gr, uprod_cum, gvabusr_adj, gdibusr_adj, aprod_log, aprod_gr, and aprod_cum from the public and adj_agg datasets.

The function returns a prod DataFrame, which contains the results of these calculations, necessary for plotting the seventh graph.

Returns

prod::DataFrame: A DataFrame that contains computed variables for Plot 7.

Offshore_Profit_Shifting_Replication.makedataforplot8 — Method

makedataforplot8()

Processes and merges several data files for use in plotting. The function reads, renames, and processes data from specified Excel and CSV files, and calculates additional columns for each industry group.

Returns

inds: A DataFrame containing processed data for each industry group (rd, nrd, itu, nitu, itp, nitp) and year, with additional computed growth rates and cumulative growth rates.

Offshore_Profit_Shifting_Replication.makeplotdata — Method

makeplotdata()

This function prepares and processes the data that will be used to create specific plots. It reads data from multiple sources, performs interpolations to handle missing values, and applies specific transformations.

The function performs the following operations:

Reads data from "aggregate.csv" and "OutputAdjNet.xlsx" files.

Applies a filter on the years between 1982 and 2017.

Interpolates missing values within the data.

Adds new columns to the DataFrame adj_agg, each being a product of an existing column and the income_adj_factor column from the public DataFrame.

Reads data from "OutputAdjNetIndustry.xlsx" and "OutputAdjNetHaven.xlsx", creates datasets with all combinations of unique years and respective unique identifier (IEDindPar and Haven), then joins them with the original data.

Filters out specific industry codes from the adj_ind and adj_tax DataFrames.

Interpolates missing values within the adj_ind and adj_tax DataFrames.

At the end, it returns four DataFrames: public, adj_agg, adj_ind, and adj_tax.

Usage

```julia public, adjagg, adjind, adjtax = makeplotdata(public, adjagg, adjind, adjtax)

Returns

public: DataFrame read from the "aggregate.csv" file.

adj_agg: DataFrame read and processed from the "OutputAdjNet.xlsx" file.

adj_ind: DataFrame read and processed from the "OutputAdjNetIndustry.xlsx" file.

adj_tax: DataFrame read and processed from the "OutputAdjNetHaven.xlsx" file.

Offshore_Profit_Shifting_Replication.plots1 — Method

plots1()

Creates a two-panel figure which shows the share of business-sector value added as a percentage for USDIA Income and Weighted Adjustment across years in panel A and their respective dollar values in panel B.

Each panel represents a distinct subplot:

Panel A: Displays the USDIA Income and Weighted Adjustment as a percentage of business-sector value added from 1980 to 2018. The USDIA Income is represented by a black line, while the Weighted Adjustment is depicted in blue.

Panel B: Shows the USDIA Income and Weighted Adjustment in billions of dollars from 1980 to 2018. Similar to panel A, the USDIA Income is shown as a black line, and the Weighted Adjustment is shown as a blue line.

The resulting figure is saved as a PDF in the folder "ReplicationFiles/figures" with the name "Figure1.pdf".

This function employs the GR backend for generating plots and uses the Plots.jl package for constructing the multi-panel figure.

Usage

```julia plots1()

No arguments are needed for this function.

Offshore_Profit_Shifting_Replication.plots10 — Method

This function replicates a figure 10, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file.

# Usage

```julia plots10()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots11 — Method

This function replicates a figure 11, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file.

# Usage

```julia plots11()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots2 — Method

plots2()

Generates a two-panel figure which displays the USDIA Income and its Weighted Adjustment, along with the Compensation, PPE, and R&D adjustments in billions of dollars from 1980 to 2018.

Each panel represents a distinct subplot:

Panel A: Illustrates the USDIA Income, Weighted Adjustment, Compensation, PPE, and R&D adjustments over the specified period. The USDIA Income is shown as a black line, the Weighted Adjustment is depicted in blue, the Compensation adjustment in a dashed red line, the PPE adjustment in a dash-dotted black line, and the R&D adjustment in a dotted green line.

Panel B: Displays the USDIA Income and Weighted Adjustment, with additional statistical information like +/-2 standard deviations range shown in light blue, and the 25th and 75th percentiles shown with dashed black lines.

The figure is saved as a PDF in the folder "ReplicationFiles/figures" with the name "Figure2.pdf".

This function uses the GR backend for generating plots and the Plots.jl package for constructing the multi-panel figure.

Usage

```julia plots2()

No arguments needed for this function

Offshore_Profit_Shifting_Replication.plots3 — Method

plots3()

Generates a two-panel figure, each panel representing a distinct subplot related to the petroleum industry and its impact on the business-sector value.

Panel A: Displays the petroleum prices over the years from 1980 to 2018. The petroleum prices are depicted as a black line in dollars per barrel.

Panel B: Illustrates two distinct measures over the same period: The Weighted Adjustment of all industries (blue line) and the adjustment excluding petroleum industry (red dashed line). Both are represented as a share of the business-sector value added, in percent.

The figure is saved as a PDF in the folder "ReplicationFiles/figures" with the name "Figure3.pdf".

This function uses the GR backend for generating plots and the Plots.jl package for constructing the multi-panel figure.

Usage

```julia plots3()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots4 — Method

This function replicates a figure 4, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file.

# Usage
```julia
plots4()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots6 — Method

This function replicates a figure 6, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file. # Usage ```julia plots6()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots7 — Method

This function replicates a figure 7, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file. # Usage ```julia plots7()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots8 — Method

This function replicates a figure 8, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file.

# Usage

```julia plots8()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.plots9 — Method

This function replicates a figure 9, which consists of two subplots. The figure is saved in the folder "ReplicationFiles/figures" as a pdf file.

# Usage

```julia plots9()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.process_files — Method

process_files(filenames)

Processes a collection of Excel files and sheets into DataFrames, storing each in a dictionary with the filename as the key.

Arguments

filenames: A collection of tuples where each tuple contains a filename and a sheet name.

Returns

A dictionary with filenames as keys and processed DataFrames as values.

Offshore_Profit_Shifting_Replication.rename_and_process! — Method

rename_and_process!(df, file)

Renames columns, interpolates missing values, and reshapes the provided DataFrame df according to certain rules based on the provided file string. The DataFrame is also sorted and certain columns are selected based on their relevance to the filename.

Arguments

df: A DataFrame to be processed.

file: A string that determines certain renaming and selection rules.

Returns

A processed DataFrame.

Offshore_Profit_Shifting_Replication.table1 — Method

This function replicates table 1. The Table is saved in the folder "ReplicationFiles/tables" as a markdown file.

# Usage

```julia table1()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.table5 — Method

This function replicates table 5. The Table is saved in the folder "ReplicationFiles/tables" as a markdown file. ```julia table5()

No arguments are needed for this function

Offshore_Profit_Shifting_Replication.table6 — Method

This function replicates table 6. The Table is saved in the folder "ReplicationFiles/tables" as a markdown file. ```julia table6()

No arguments are needed for this function